updating yamldocs and stencil to support template fallback
